requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
numpy==1.23.3
tensorflow-cpu==2.10.0
python-dotenv==0.21.0
boto3==1.26.2
Pillow==9.3.0
requests==2.28.1
gunicorn==20.1.0
h5py==3.7.0